+++ emulator/src/ro/dobrescuandrei/yaktnes/cpu/instruction/execution/strategy/InstructionWithDecimalArgumentExecutionStrategy.kt
@@ -1,15 +1,15 @@
 package ro.dobrescuandrei.yaktnes.cpu.instruction.execution.strategy
 
-import ro.dobrescuandrei.yaktnes.cpu.datatype.Decimal
+import ro.dobrescuandrei.yaktnes.cpu.datatype.Int8
 import ro.dobrescuandrei.yaktnes.cpu.instruction.addressing_mode.InstructionArgumentFactory
 import ro.dobrescuandrei.yaktnes.cpu.instruction.definition.InstructionDefinition
-import ro.dobrescuandrei.yaktnes.cpu.machine_code.MachineCode
+import ro.dobrescuandrei.yaktnes.cpu.MachineCode
 
-class InstructionWithDecimalArgumentExecutionStrategy : InstructionExecutionStrategy<Decimal>
+class InstructionWithDecimalArgumentExecutionStrategy : InstructionExecutionStrategy<Int8>
 {
-    override fun getInstructionArgumentType() = Decimal::class.java
+    override fun getInstructionArgumentType() = Int8::class.java
 
-    override fun executeInstruction(definition : InstructionDefinition<Decimal>, machineCode : MachineCode)
+    override fun executeInstruction(definition : InstructionDefinition<Int8>, machineCode : MachineCode)
     {
         val value=InstructionArgumentFactory.getDecimalArgument(
                 addressingMode = definition.addressingMode,
